requirements.txt
openai>=1.1.0
pyzotero>=1.5.5
pydub>=0.25.1
podgen>=1.1.0
boto3>=1.26.0
requests>=2.28.1
python-dotenv>=0.21.0 